Define flexlm servers here.
 It's referenced via windows system environment variable called "toolboxxml". In the functions ps1, the path is defined as $env:toolboxxml + "\etc\flexlm_monitor.xml" so you'll need to store that somewhere and set up the variable, or change it around to your liking.

Define each daemon as a "service" running on a particular host. (hosts can run multiple flexlm services, on different tcp ports)

The regexlibrary is separated from the functions, to facilitate readability, and allow addition filters or improvements to be made. You can see those references in the ps1 file.

You also need to take care to include lmutil.exe, which is required to communicate with the flexlm service. I am basically doing a fancy "screen scrape" of lmutil output. It would be nice to take a direct TCP socket approach, but there appears to be some handshake that happens that I'm not patient enough to work out.