simplified config file reading
This code didn't really work "out of the box" because it was intended to live with a "toolbox" framework and some dependencies. Just make sure that "flexlm_monitor.xml" configured appropriately is in the current path you're executing. You should be able to get the gist of how this works and modify to suit how you want to save the config file.